recode and forgive
as I was using this function in another package I got annoyed at the errors I got when a factor variable had levels but not labels, particularly in the case where I didnt feel like I needed to write labels bc the levels were just as good. This will make situations like that throw a warning but not an error.